Ensure stall_load doesn't incorrectly activate
When i-cache responds, stall_load can incorrectly activate if the mem
stage is also doing a mem operation and the d-cache gets a hit. In this
case an instruction gets incorrectly deleted since stall_load activates
incorrectly, causing incorrect behaviour. This causes stall_load to not
active if we are waiting on an i-cache operation.